Barebones guzzle adapter for flysystem v3
